requirements.txt
torch>=2.0.0
diffusers==0.32.2
transformers==4.49.0
accelerate==1.4.0
safetensors>=0.3.1
pillow>=9.0.0
huggingface-hub>=0.23.2
imageio>=2.31.0
imageio-ffmpeg>=0.4.8 